refactor: Remove "TestingTreeSingleton" class
