+++ src/modules/testing-tree/index.ts
@@ -1,4 +1,3 @@
-export { TestingTree } from "./store/testing-tree";
-export { TestingTreeSingleton } from "./store/testing-tree-singleton";
+export { TestingTree } from "./model/testing-tree";
 export { GroupNode } from "./classes/group-node";
 export { type TestNode } from "./types";
