--- src/modules/testing-tree/store/testing-tree-singleton.ts
@@ -1,19 +1,0 @@
-import { TestingTree } from "@modules/testing-tree";
-
-export class TestingTreeSingleton {
-    public tree: TestingTree;
-
-    private static instance: TestingTreeSingleton | null = null;
-
-    private constructor() {
-        this.tree = new TestingTree();
-    }
-
-    public static getInstance() {
-        if (!TestingTreeSingleton.instance) {
-            TestingTreeSingleton.instance = new TestingTreeSingleton();
-        }
-
-        return TestingTreeSingleton.instance;
-    }
-}
